add `worker.py` & use a workaround to support the df merging
